+++ app/common/src/commonMain/kotlin/com/seiko/imageloader/demo/scene/Common.kt
@@ -24,7 +24,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
-import com.seiko.imageloader.compose.AutoSizeBox
+import com.seiko.imageloader.ui.AutoSizeBox
 import com.seiko.imageloader.demo.model.Image
 import com.seiko.imageloader.demo.util.NullDataInterceptor
 import com.seiko.imageloader.demo.util.decodeJson
