--- image-loader/src/commonMain/kotlin/com/seiko/imageloader/compose/CachedPositionAndSize.kt
@@ -1,9 +1,0 @@
-package com.seiko.imageloader.compose
-
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
-
-internal data class CachedPositionAndSize(
-    val position: Offset,
-    val size: Size,
-)
