+++ image-loader/src/commonMain/kotlin/com/seiko/imageloader/ui/AutoSizeImage.kt
@@ -1,4 +1,4 @@
-package com.seiko.imageloader.compose
+package com.seiko.imageloader.ui
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.RememberObserver
